+++ src/Enums/HelperGenerateLocaleEnum.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atlcom\Enums;
+
+/**
+ * Локаль
+ */
+enum HelperGenerateLocaleEnum: string
+{
+    case English = 'english';
+    case Russian = 'russian';
+}