+++ src/Enums/HelperStringBreakTypeEnum.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Atlcom\Enums;
 
+/**
+ * Разделение текста
+ */
 enum HelperStringBreakTypeEnum: string
 {
     case Word = 'word';