+++ src/Enums/HelperStringEncodingEnum.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Atlcom\Enums;
 
+/**
+ * Кодировка
+ */
 enum HelperStringEncodingEnum: string
 {
     case Utf8 = 'UTF-8';