+++ src/Enums/HelperTranslitDirectionEnum.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Atlcom\Enums;
 
+/**
+ * Направление транслитерации
+ */
 enum HelperTranslitDirectionEnum: string
 {
     case Auto = 'auto';