+++ src/Traits/HelperGenerateTrait.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Atlcom\Traits;
 
+use Atlcom\Consts\HelperConsts as Consts;
+use Atlcom\Enums\HelperGenerateLocaleEnum as Locale;
+
 /**
  * Трейт для работы с генерацией данных
  * @mixin \Atlcom\Helper
@@ -54,7 +57,7 @@ trait HelperGenerateTrait
         !(is_string($paths) && static::regexpValidatePattern($paths))
             ?: $paths = [static::stringRandom($paths)];
         $paths ??= [static::stringRandom('/[a-z0-9\-\/]{0,20}/')];
-        $path = $paths[array_rand($paths)];
+        $path = $paths[array_rand($paths)] ?: '/';
 
         !(is_string($queries) && static::regexpValidatePattern($queries))
             ?: $queries = [static::stringRandom($queries)];
@@ -89,13 +92,15 @@ trait HelperGenerateTrait
     }
 
 
-    //?!? test
+    //?!? readme
     /**
-     * Генерация случайного пароля
+     * Возвращает случайно сгенерированный пароль
+     * @see ../../tests/HelperGenerateTrait/HelperGeneratePasswordTest.php
+     * 
      * @param string $pattern
      * @return string
      */
-    public static function generatePassword(string $pattern = '/[A-Za-z0-9!@#$%^&*()_+]{8,16}/'): string
+    public static function generatePassword(string $pattern = '/[A-Za-z0-9!@#$%^&\*()_\+]{8,16}/'): string
     {
         return static::stringRandom($pattern);
     }
@@ -103,7 +108,9 @@ trait HelperGenerateTrait
 
     //?!? test
     /**
-     * Генерация случайного email
+     * Возвращает случайно сгенерированный email
+     * @see ../../tests/HelperGenerateTrait/HelperGenerateEmailTest.php
+     * 
      * @return string
      */
     public static function generateEmail(): string
@@ -118,7 +125,9 @@ trait HelperGenerateTrait
 
     //?!? test
     /**
-     * Генерация случайного телефона с кодом страны
+     * Возвращает случайно сгенерированный номер телефона с кодом страны
+     * @see ../../tests/HelperGenerateTrait/HelperGeneratePhoneTest.php
+     * 
      * @param string $countryCode
      * @return string
      */
@@ -133,10 +142,12 @@ trait HelperGenerateTrait
 
     //?!? test
     /**
-     * Генерация случайного UUID v4
+     * Возвращает случайно сгенерированный UUID v4
+     * @see ../../tests/HelperGenerateTrait/HelperGenerateUuid4Test.php
+     * 
      * @return string
      */
-    public static function generateUuid(): string
+    public static function generateUuid4(): string
     {
         $data = random_bytes(16);
         $data[6] = chr((ord($data[6]) & 0x0f) | 0x40); // version 4
@@ -146,9 +157,98 @@ trait HelperGenerateTrait
     }
 
 
-    //?!? code
-    public static function generateName(): string
+    //?!? test
+    /**
+     * Возвращает случайно сгенерированный UUID v7
+     * @see ../../tests/HelperGenerateTrait/HelperGenerateUuid7Test.php
+     * 
+     * @return string
+     */
+    public static function generateUuid7(): string
     {
-        return '';
+        // Uuidv7: 48 бит Unix TimeStamp MS, 12 бит случайные, 2 -битные версии, 62 бита случайные
+        $time = (int) (microtime(true) * 1000); // миллисекунды
+        $timeHex = str_pad(dechex($time), 12, '0', STR_PAD_LEFT);
+
+        $rand = bin2hex(random_bytes(10)); // 20 шестнадцатеричных частей = 80 бит
+
+        // Формируем UUIDv7: xxxxxxxx-xxxx-7xxx-yxxx-xxxxxxxxxxxx
+        $uuid = sprintf(
+            '%08s-%04s-7%03s-%s%s-%s',
+            substr($timeHex, 0, 8),
+            substr($timeHex, 8, 4),
+            substr($rand, 0, 3),
+            dechex((hexdec(substr($rand, 3, 1)) & 0x3) | 0x8), // вариант 10xx
+            substr($rand, 4, 3),
+            substr($rand, 7)
+        );
+
+        return $uuid;
+    }
+
+
+    //?!? test
+    /**
+     * Возвращает случайно сгенерированное ФИО
+     * @see ../../tests/HelperGenerateTrait/HelperGenerateNameTest.php
+     *
+     * @param Locale $locale
+     * @param array|string|bool|null $surnames
+     * @param array|string|bool|null $firstNames
+     * @param array|string|bool|null $patronymics
+     * @return string
+     */
+    public static function generateName(
+        Locale $locale = Locale::Russian,
+        array|string|bool|null $surnames = true,
+        array|string|bool|null $firstNames = true,
+        array|string|bool|null $patronymics = false,
+    ): string {
+        $surnamePrefixes = in_array($surnames, [null, false], true)
+            ? match ($locale) {
+                Locale::English => Consts::SURNAME_PREFIXES_EN,
+                Locale::Russian => Consts::SURNAME_PREFIXES_RU,
+            } : [];
+        $surnameSuffixes = in_array($surnames, [null, false], true) ? match ($locale) {
+            Locale::English => Consts::SURNAME_SUFFIXES_EN,
+            Locale::Russian => Consts::SURNAME_SUFFIXES_RU,
+        } : [];
+        $firstNamePrefixes = in_array($firstNames, [null, false], true)
+            ? match ($locale) {
+                Locale::English => Consts::FIRSTNAME_PREFIXES_EN,
+                Locale::Russian => Consts::FIRSTNAME_PREFIXES_RU,
+            } : [];
+        $firstNameSuffixes = in_array($firstNames, [null, false], true)
+            ? match ($locale) {
+                Locale::English => Consts::FIRSTNAME_PREFIXES_EN,
+                Locale::Russian => Consts::FIRSTNAME_PREFIXES_RU,
+            } : [];
+        $patronymicPrefixes = in_array($patronymics, [null, false], true)
+            ? match ($locale) {
+                Locale::English => [],
+                Locale::Russian => Consts::PATRONYNIC_PREFIXES_RU,
+            } : [];
+        $patronymicSuffixes = in_array($patronymics, [null, false], true)
+            ? match ($locale) {
+                Locale::English => [],
+                Locale::Russian => Consts::PATRONYNIC_SUFFIXES_RU,
+            } : [];
+
+        $names = [];
+
+        if ($surnamePrefixes) {
+            $names[] = $surnamePrefixes[array_rand($surnamePrefixes)]
+                . ($surnameSuffixes ? $surnameSuffixes[array_rand($surnameSuffixes)] : '');
+        }
+        if ($firstNamePrefixes) {
+            $names[] = $firstNamePrefixes[array_rand($firstNamePrefixes)]
+                . ($firstNameSuffixes ? $firstNameSuffixes[array_rand($firstNameSuffixes)] : '');
+        }
+        if ($patronymicPrefixes) {
+            $names[] = $patronymicPrefixes[array_rand($patronymicPrefixes)]
+                . ($patronymicSuffixes ? $patronymicSuffixes[array_rand($patronymicSuffixes)] : '');
+        }
+
+        return implode(' ', $names);
     }
 }