+++ tests/HelperGenerateTrait/HelperGenerateUuid7Test.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atlcom\Tests\HelperGenerateTrait;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Тест метода трейта
+ * @see \Atlcom\Traits\HelperGenerateTrait
+ */
+final class HelperGenerateUuid7Test extends TestCase
+{
+}